Adds UI support for the General category of the GBIF Extended Metadata Profile (see definition of elements document: http://goo.gl/Cjan). Note that the form submission code has not yet been updated to receive the new values on the server side. This will be committed separately.
Details:

The 'Responsible Party' (item 7 in the General category) in the UI remains on the Resource Originator page and was not moved into the Basic Metadata page. Also, the 'Keywords' and 'Keyword Thesaurus' (items 8 and 9, respectively, in the General category) in the UI remain in the General Keywords page. 

Role and MetadataAction classes were modified to support displaying a UI select box of internationalizable role types. English translations were added to ApplicationResources.properties file.