This commit is a nearly complete attempt to display publishing status in Manage/Resources resource list table as in the Metadata Forms resource publication status box. Everything is in place to use private static class ResourceForDisplay to render status as internationalised private/unpublished changes/published. Note that PublicationStatus enum can't be changed without drastic refactoring even into H2. The resourceManagerList.ftl shows how the MetadataAction.resouresForDisplay list is meant to be used, but is is not currently functional due to exceptions trying to invoke getters for the ResourceDisplay properties. I have run out of time trying to figure out why the getters aren't being seen. To recover previous functionality for this resourceManagerList table, only lines 4 and 7 in resourceManagerList.ftl have to be reverted to the previous revision.
Once the getter problem is solved, the UI should show "private" for the status of an unpublished resource, "unpublished changes" for the status of a resource that has been published, but not since changes were made, and "published" if the resource has been published and no further changes have been made since. This should satisfy the simplification of the publishing UI, but does not address the publishing work flow, as agreed we would not as part of the 1.1 deliverables, but which should be seriously considered for the next round of changes.